Introduce SystemTableProvider
