+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/HiveMetadataFactory.java
@@ -29,6 +29,7 @@ import io.trino.spi.type.TypeManager;
 import javax.inject.Inject;
 
 import java.util.Optional;
+import java.util.Set;
 import java.util.concurrent.Executor;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.ScheduledExecutorService;
@@ -59,6 +60,7 @@ public class HiveMetadataFactory
     private final Executor updateExecutor;
     private final String trinoVersion;
     private final HiveRedirectionsProvider hiveRedirectionsProvider;
+    private final Set<SystemTableProvider> systemTableProviders;
     private final HiveMaterializedViewMetadataFactory hiveMaterializedViewMetadataFactory;
     private final AccessControlMetadataFactory accessControlMetadataFactory;
     private final Optional<Duration> hiveTransactionHeartbeatInterval;
@@ -79,6 +81,7 @@ public class HiveMetadataFactory
             JsonCodec<PartitionUpdate> partitionUpdateCodec,
             NodeVersion nodeVersion,
             HiveRedirectionsProvider hiveRedirectionsProvider,
+            Set<SystemTableProvider> systemTableProviders,
             HiveMaterializedViewMetadataFactory hiveMaterializedViewMetadataFactory,
             AccessControlMetadataFactory accessControlMetadataFactory)
     {
@@ -105,6 +108,7 @@ public class HiveMetadataFactory
                 heartbeatService,
                 nodeVersion.toString(),
                 hiveRedirectionsProvider,
+                systemTableProviders,
                 hiveMaterializedViewMetadataFactory,
                 accessControlMetadataFactory);
     }
@@ -132,6 +136,7 @@ public class HiveMetadataFactory
             ScheduledExecutorService heartbeatService,
             String trinoVersion,
             HiveRedirectionsProvider hiveRedirectionsProvider,
+            Set<SystemTableProvider> systemTableProviders,
             HiveMaterializedViewMetadataFactory hiveMaterializedViewMetadataFactory,
             AccessControlMetadataFactory accessControlMetadataFactory)
     {
@@ -152,6 +157,7 @@ public class HiveMetadataFactory
         this.partitionUpdateCodec = requireNonNull(partitionUpdateCodec, "partitionUpdateCodec is null");
         this.trinoVersion = requireNonNull(trinoVersion, "trinoVersion is null");
         this.hiveRedirectionsProvider = requireNonNull(hiveRedirectionsProvider, "hiveRedirectionsProvider is null");
+        this.systemTableProviders = requireNonNull(systemTableProviders, "systemTableProviders is null");
         this.hiveMaterializedViewMetadataFactory = requireNonNull(hiveMaterializedViewMetadataFactory, "hiveMaterializedViewMetadataFactory is null");
         this.accessControlMetadataFactory = requireNonNull(accessControlMetadataFactory, "accessControlMetadataFactory is null");
         this.hiveTransactionHeartbeatInterval = requireNonNull(hiveTransactionHeartbeatInterval, "hiveTransactionHeartbeatInterval is null");
@@ -199,6 +205,7 @@ public class HiveMetadataFactory
                 trinoVersion,
                 new MetastoreHiveStatisticsProvider(metastore),
                 hiveRedirectionsProvider,
+                systemTableProviders,
                 hiveMaterializedViewMetadataFactory.create(hiveMetastoreClosure),
                 accessControlMetadataFactory.create(metastore));
     }
@@ -218,6 +225,7 @@ public class HiveMetadataFactory
             String trinoVersion,
             HiveStatisticsProvider hiveStatisticsProvider,
             HiveRedirectionsProvider hiveRedirectionsProvider,
+            Set<SystemTableProvider> systemTableProviders,
             HiveMaterializedViewMetadata hiveMaterializedViewMetadata,
             AccessControlMetadata accessControlMetadata)
     {
@@ -236,6 +244,7 @@ public class HiveMetadataFactory
                 trinoVersion,
                 hiveStatisticsProvider,
                 hiveRedirectionsProvider,
+                systemTableProviders,
                 hiveMaterializedViewMetadata,
                 accessControlMetadata);
     }
