+++ plugin/trino-hive/src/test/java/io/trino/plugin/hive/AbstractTestHive.java
@@ -829,6 +829,9 @@ public abstract class AbstractTestHive
                             ImmutableMap.of(),
                             TupleDomain.all()));
                 },
+                ImmutableSet.of(
+                        new PartitionsSystemTableProvider(partitionManager),
+                        new PropertiesSystemTableProvider()),
                 (metastore) -> new NoneHiveMaterializedViewMetadata()
                 {
                     @Override
